requirements.txt
djangorestframework
Pillow==8.3.2
psycopg2-binary